Struct axnet::TcpSocket

pub struct TcpSocket { /* private fields */ }

A TCP socket that provides POSIX-like APIs.

connect is for TCP clients.

bind, listen, and accept are for TCP servers.

Other methods are for both TCP clients and servers.

Implementations

impl TcpSocket

pub const fn new() -> Self

Creates a new TCP socket.

pub fn local_addr(&self) -> AxResult<SocketAddr>

Returns the local address and port, or Err(NotConnected) if not connected.

pub fn peer_addr(&self) -> AxResult<SocketAddr>

Returns the remote address and port, or Err(NotConnected) if not connected.

pub fn is_nonblocking(&self) -> bool

Returns whether this socket is in nonblocking mode.

pub fn set_nonblocking(&self, nonblocking: bool)

Moves this TCP stream into or out of nonblocking mode.

This will result in read, write, recv and send operations becoming nonblocking, i.e., immediately returning from their calls. If the IO operation is successful, Ok is returned and no further action is required. If the IO operation could not be completed and needs to be retried, an error with kind Err(WouldBlock) is returned.

pub fn connect(&self, remote_addr: SocketAddr) -> AxResult

Connects to the given address and port.

The local port is generated automatically.

pub fn bind(&self, local_addr: SocketAddr) -> AxResult

Binds an unbound socket to the given address and port.

If the given port is 0, it generates one automatically.

It’s must be called before listen and accept.

pub fn listen(&self) -> AxResult

Starts listening on the bound address and port.

It’s must be called after bind and before accept.

pub fn accept(&self) -> AxResult<TcpSocket>

Accepts a new connection.

This function will block the calling thread until a new TCP connection is established. When established, a new TcpSocket is returned.

It’s must be called after bind and listen.

pub fn shutdown(&self) -> AxResult

Close the connection.

pub fn close(&mut self)

Close the transmit half of the tcp socket. It will call close() on smoltcp::socket::tcp::Socket. It should send FIN to remote half.

This function is for shutdown(fd, SHUT_WR) syscall.

It won’t change TCP state. It won’t affect unconnected sockets (listener).

pub fn recv(&self, buf: &mut [u8]) -> AxResult<usize>

Receives data from the socket, stores it in the given buffer.

pub fn recv_timeout(&self, buf: &mut [u8], ticks: u64) -> AxResult<usize>

Receives data from the socket, stores it in the given buffer.

It will return Err(Timeout) if expired.

pub fn send(&self, buf: &[u8]) -> AxResult<usize>

Transmits data in the given buffer.

pub fn poll(&self) -> AxResult<PollState>

Whether the socket is readable or writable.

pub fn set_nagle_enabled(&self, enabled: bool) -> AxResult

To set the nagle algorithm enabled or not.

pub fn nagle_enabled(&self) -> bool

To get the nagle algorithm enabled or not.

pub fn with_socket<R>(&self, f: impl FnOnce(Option<&Socket<'_>>) -> R) -> R

To get the socket and call the given function.

If the socket is not connected, it will return None.

Or it will return the result of the given function.

pub fn with_socket_mut<R>( &mut self, f: impl FnOnce(Option<&mut Socket<'_>>) -> R ) -> R

To get the mutable socket and call the given function.

If the socket is not connected, it will return None.

Or it will return the result of the given function.

impl TcpSocket

Private methods

pub fn is_connected(&self) -> bool

Whether the socket is connected.