axnet

Struct UdpSocket

pub struct UdpSocket { /* private fields */ }

A UDP socket that provides POSIX-like APIs.

Implementations

impl UdpSocket

pub fn new() -> Self

Creates a new UDP socket.

pub fn local_addr(&self) -> AxResult<SocketAddr>

Returns the local address and port, or Err(NotConnected) if not connected.

pub fn peer_addr(&self) -> AxResult<SocketAddr>

Returns the remote address and port, or Err(NotConnected) if not connected.

pub fn is_nonblocking(&self) -> bool

Returns whether this socket is in nonblocking mode.

pub fn set_nonblocking(&self, nonblocking: bool)

Moves this UDP socket into or out of nonblocking mode.

This will result in recv, recv_from, send, and send_to operations becoming nonblocking, i.e., immediately returning from their calls. If the IO operation is successful, Ok is returned and no further action is required. If the IO operation could not be completed and needs to be retried, an error with kind Err(WouldBlock) is returned.

pub fn bind(&self, local_addr: SocketAddr) -> AxResult

Binds an unbound socket to the given address and port.

It’s must be called before send_to and recv_from.

pub fn send_to(&self, buf: &[u8], remote_addr: SocketAddr) -> AxResult<usize>

Sends data on the socket to the given address. On success, returns the number of bytes written.

pub fn recv_from(&self, buf: &mut [u8]) -> AxResult<(usize, SocketAddr)>

Receives a single datagram message on the socket. On success, returns the number of bytes read and the origin.

pub fn peek_from(&self, buf: &mut [u8]) -> AxResult<(usize, SocketAddr)>

Receives a single datagram message on the socket, without removing it from the queue. On success, returns the number of bytes read and the origin.

pub fn connect(&self, addr: SocketAddr) -> AxResult

Connects this UDP socket to a remote address, allowing the send and recv to be used to send data and also applies filters to only receive data from the specified address.

The local port will be generated automatically if the socket is not bound. It’s must be called before send and recv.

pub fn send(&self, buf: &[u8]) -> AxResult<usize>

Sends data on the socket to the remote address to which it is connected.

pub fn recv(&self, buf: &mut [u8]) -> AxResult<usize>

Receives a single datagram message on the socket from the remote address to which it is connected. On success, returns the number of bytes read.

pub fn shutdown(&self) -> AxResult

Close the socket.

pub fn poll(&self) -> AxResult<PollState>

Whether the socket is readable or writable.