axnet

Struct TcpSocket

pub struct TcpSocket { /* private fields */ }

A TCP socket that provides POSIX-like APIs.

connect is for TCP clients.

bind, listen, and accept are for TCP servers.

Other methods are for both TCP clients and servers.

Implementations

impl TcpSocket

pub const fn new() -> Self

Creates a new TCP socket.

pub fn local_addr(&self) -> AxResult<SocketAddr>

Returns the local address and port, or Err(NotConnected) if not connected.

pub fn peer_addr(&self) -> AxResult<SocketAddr>

Returns the remote address and port, or Err(NotConnected) if not connected.

pub fn is_nonblocking(&self) -> bool

Returns whether this socket is in nonblocking mode.

pub fn set_nonblocking(&self, nonblocking: bool)

Moves this TCP stream into or out of nonblocking mode.

This will result in read, write, recv and send operations becoming nonblocking, i.e., immediately returning from their calls. If the IO operation is successful, Ok is returned and no further action is required. If the IO operation could not be completed and needs to be retried, an error with kind Err(WouldBlock) is returned.

pub fn connect(&self, remote_addr: SocketAddr) -> AxResult

Connects to the given address and port.

The local port is generated automatically.

pub fn bind(&self, local_addr: SocketAddr) -> AxResult

Binds an unbound socket to the given address and port.

If the given port is 0, it generates one automatically.

It’s must be called before listen and accept.

pub fn listen(&self) -> AxResult

Starts listening on the bound address and port.

It’s must be called after bind and before accept.

pub fn accept(&self) -> AxResult<TcpSocket>

Accepts a new connection.

This function will block the calling thread until a new TCP connection is established. When established, a new TcpSocket is returned.

It’s must be called after bind and listen.

pub fn shutdown(&self) -> AxResult

Close the connection.

pub fn recv(&self, buf: &mut [u8]) -> AxResult<usize>

Receives data from the socket, stores it in the given buffer.

pub fn send(&self, buf: &[u8]) -> AxResult<usize>

Transmits data in the given buffer.

pub fn poll(&self) -> AxResult<PollState>

Whether the socket is readable or writable.